+++ src/main/arrpc.ts
@@ -26,8 +26,8 @@ export async function initArRPC() {
 
             await sendRendererCommand(IpcCommands.RPC_INVITE, invite).then(callback);
         });
-        server.on("link", (data: any, deepCallback: (valid: boolean) => void) => {
-            mainWin.webContents.executeJavaScript(`Vesktop.openDeepLink(${JSON.stringify(data)})`).then(deepCallback);
+        server.on("link", async (data: any, deepCallback: (valid: boolean) => void) => {
+            await sendRendererCommand(IpcCommands.RPC_DEEP_LINK, data).then(deepCallback);
         });
     } catch (e) {
         console.error("Failed to start arRPC server", e);
