+++ src/shared/IpcEvents.ts
@@ -61,6 +61,7 @@ export const enum IpcEvents {
 export const enum IpcCommands {
     RPC_ACTIVITY = "rpc:activity",
     RPC_INVITE = "rpc:invite",
+    RPC_DEEP_LINK = "rpc:deep",
     NAVIGATE_SETTINGS = "navigate:settings",
     GET_LANGUAGES = "navigator.languages"
 }
